Move parenthesis categorization into a new class
