+++ lib/rules/require-spaces-inside-parenthesized-expression.js
@@ -41,7 +41,7 @@
  */
 
 var assert = require('assert');
-var utils = require('../utils');
+var TokenCategorizer = require('../token-categorizer');
 
 module.exports = function() {};
 
@@ -83,7 +83,7 @@ module.exports.prototype = {
             }
 
             // Skip non-expression parentheses
-            var type = utils.categorizeOpenParen(token, file);
+            var type = TokenCategorizer.categorizeOpenParen(token, file);
             if (type !== 'ParenthesizedExpression') {
                 return;
             }
@@ -105,7 +105,7 @@ module.exports.prototype = {
             }
 
             // Skip non-expression parentheses
-            var type = utils.categorizeCloseParen(token, file);
+            var type = TokenCategorizer.categorizeCloseParen(token, file);
             if (type !== 'ParenthesizedExpression') {
                 return;
             }
